Improved readme.md

+++ evaluator/languages.py
@@ -4,15 +4,11 @@ Created on 8.6.2014
 @author: Stefan Smihla
 '''
 
-import configparser
 import os, subprocess, signal
-import platform, shutil
+import platform
 import inspect
 from evaluator import config
 from threading import Thread
-
-if platform.system() == 'Windows':
-    import win32api
 
 SOURCE_DIR = 'sources'
 
@@ -20,10 +16,12 @@ def load_languages():
     """ Loads languages from configuration file. """
 
     languages = []
-    for name, lang in sorted(inspect.getmembers(config, inspect.isclass), key=lambda x: x[1].ORDER):
+    for name, lang in sorted(inspect.getmembers(config, inspect.isclass),
+                             key=lambda x: x[1].ORDER):
+
         if name not in ['CompiledLanguage', 'Language']:
             lang_class = globals().get(name)
-            languages.append(lang_class(lang))    
+            languages.append(lang_class(lang))
 
     return languages
 
@@ -54,34 +52,32 @@ class Language(object):
         return cls._process
 
     @classmethod
+    def build_proc_params(cls):
+        params = {'universal_newlines': True,
+                  'stderr': subprocess.STDOUT,
+                  'stdout': subprocess.PIPE,
+                  'shell': True}
+
+        if platform.system() != 'Windows':
+            params['preexec_fn'] = os.setsid
+
+        return params
+
+    @classmethod
     def run_process(cls, command):
         """ Run process and saves it output. """
-        cls.set_process(subprocess.Popen(command,
-                                        universal_newlines=True,
-                                        stderr=subprocess.STDOUT,
-                                        stdout=subprocess.PIPE,
-                                        shell=True,
-                                        preexec_fn=os.setsid))
+        cls.set_process(subprocess.Popen(command, **cls.build_proc_params()))
         cls.set_output(cls.get_process().stdout.read().strip())
         return cls.get_stdout()
-        
-    def safe_output(self, output=None):
-        if not output:
-            output = self.output
-
-        if os.path.dirname(output):
-            return output
-        else:
-            return os.path.join('.', output)
 
     def __init__(self, config_lang):
         self.name = config_lang.NAME
         self.program = config_lang.PROGRAM
-        self.available = shutil.which(self.program)
-        
+        self.available = True if self.program else False
+
         if not self.available:
             return
-        
+
         self.version_cmd = config_lang.VERSION
         self.compile_cmd = config_lang.COMPILE
         self.run_cmd = config_lang.RUN
@@ -105,7 +101,13 @@ class Language(object):
         if self.get_stdout():
             return self.get_stdout()
         else:
-            os.killpg(self.get_process().pid, signal.SIGTERM)
+            if platform.system() == 'Windows':
+                subprocess.Popen('TASKKILL /F /PID %s /T' \
+                                 % self.get_process().pid,
+                                 stdout=subprocess.DEVNULL,
+                                 stderr=subprocess.DEVNULL)
+            else:
+                os.killpg(self.get_process().pid, signal.SIGTERM)
             self.set_process(None)
             return '%s:\nKilled' % self.name
 
@@ -154,10 +156,6 @@ class JavaScriptLanguage(Language):
 
 class PHPLanguage(Language):
     """ PHP Language class. """
-    pass
-
-class RubyLanguage(Language):
-    """ Ruby Language class. """
     pass
 
 class PerlLanguage(Language):
